+++ dev/towerdefense/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 1.2 - dev
 
 - CHANGE: Change update `updateHealthHud` to `updateHUD`. `updateHUD` will update every hud gold and health
+- CHANGE: Now if there's no `Plugin Command Config` in the map, the game will not create hud or any tower defense mechanic.
+- ADD: `Plugin Command: Disable Tower Defense` This will destroy any tower defense element and enable menu
 
 ## Version 1.1 - 6/12/2020